Re-implement admin hover functionality without cookies and javascript
The cookie that this functionality was relying on has not been set since
2016. This re-implements the function in the backend. This means the
resulting HTML should not be cached.

For comments and trackbacks, the helper call is simply removed, since the
edit link was also missing.

To avoid breaking any themes that rely on it, the helper itself is just
made a no-op.